requirements.txt
torch>=2.0
numpy
matplotlib
scikit-image
imageio
Pillow
# optional: interactive viewer
viser
scipy
